putting main wrapper back in agent
